Refactor code to remove origin address access

+++ ui/dlg_blockchain.go
@@ -95,7 +95,7 @@ func DlgBlockchain(name string) *gocui.Popup {
 					}
 
 					for i, bch := range cmn.CurrentWallet.Blockchains {
-						if bch.ChainId == uint(chainid) && (i == -1 || i != bch_index) {
+						if bch.ChainId == chainid && (i == -1 || i != bch_index) {
 							Notification.ShowErrorf("Chain id %d already used by %s", chainid, bch.Name)
 							break
 						}
@@ -127,7 +127,7 @@ func DlgBlockchain(name string) *gocui.Popup {
 					if bch_index != -1 {
 						cmn.CurrentWallet.Blockchains[bch_index].Name = name
 						cmn.CurrentWallet.Blockchains[bch_index].Url = rpc
-						cmn.CurrentWallet.Blockchains[bch_index].ChainId = uint(chainid)
+						cmn.CurrentWallet.Blockchains[bch_index].ChainId = chainid
 						cmn.CurrentWallet.Blockchains[bch_index].ExplorerUrl = explorer
 						cmn.CurrentWallet.Blockchains[bch_index].Currency = currency
 						cmn.CurrentWallet.Blockchains[bch_index].WTokenAddress = wta
@@ -135,7 +135,7 @@ func DlgBlockchain(name string) *gocui.Popup {
 						cmn.CurrentWallet.Blockchains = append(cmn.CurrentWallet.Blockchains, &cmn.Blockchain{
 							Name:          name,
 							Url:           rpc,
-							ChainId:       uint(chainid),
+							ChainId:       chainid,
 							ExplorerUrl:   explorer,
 							Currency:      currency,
 							WTokenAddress: wta,
